fix double folder problem
