+++ ChoreBuddies.Backend/Features/Chores/IChoreService.cs
@@ -1,0 +1,9 @@
+﻿using Shared.Chores;
+
+namespace ChoreBuddies.Backend.Features.Chores;
+
+public interface IChoresService
+{
+    IEnumerable<ChoreOverviewDto> GetChores();
+    ChoreDto GetChoreDetails(string id);
+}
