--- ChoreBuddies.Backend/features/Chores/IChoreService.cs
@@ -1,9 +1,0 @@
-﻿using Shared.Chores;
-
-namespace ChoreBuddies.Backend.Features.Chores;
-
-public interface IChoresService
-{
-    IEnumerable<ChoreOverviewDto> GetChores();
-    ChoreDto GetChoreDetails(string id);
-}
