--- ChoreBuddies.Backend/features/Home/HomeController.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace ChoreBuddies.Backend.Features.Home;
-
-[ApiController]
-[Route("api/v1/[controller]")]
-public class HomeController : Controller
-{
-    [HttpGet]
-    public IActionResult Weather()
-    {
-        return Ok(new HomeObject("Welcome!"));
-    }
-    protected class HomeObject
-    {
-        public string? message { get; set; }
-        public HomeObject(string? Message)
-        {
-            message = Message;
-        }
-    }
-}
